+++ src/entities/canvas_manager/base-canvas-manager.ts
