+++ src/entities/canvas_manager/offline-canvas-manager.ts
@@ -1,4 +1,4 @@
-import { BaseCanvasManager } from './base_canvas_manager'
+import { BaseCanvasManager } from './base-canvas-manager'
 import { Box } from '../box'
 import { Maguma } from '../maguma'
 import { Canvas } from '../canvas'
